# commons/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "IOTGlue NodeAPI's",
    "description": "",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/iotglueapi/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/commune/templates": {
      "get": {
        "description": "",
        "summary": "getAllTemplatesWithPagination",
        "tags": [
          "Template"
        ],
        "operationId": "getAllTemplatesWithPagination",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page_no",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "page_size",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "summary": "createTemplate",
        "tags": [
          "Template"
        ],
        "operationId": "createTemplate",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createTemplateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/commune/templates/{templateId}": {
      "delete": {
        "description": "",
        "summary": "deleteTemplateById",
        "tags": [
          "Template"
        ],
        "operationId": "deleteTemplateById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "templateId",
          "in": "path",
          "description": "ID of template to delete",
          "required": true,
          "type": "string"
          }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "",
        "summary": "getTemplateById",
        "tags": [
          "Template"
        ],
        "operationId": "getTemplateById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "templateId",
          "in": "path",
          "description": "ID of template to return",
          "required": true,
          "type": "string"
          }],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "",
        "summary": "updateTemplateById",
        "tags": [
          "Template"
        ],
        "operationId": "updateTemplateById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "templateId",
            "in": "path",
            "description": "ID of template to update",
            "required": true,
            "type": "string"
            },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateTemplateByIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/common/rightpanel/templates/{templateId}": {
      "get": {
        "description": "getTemplateForRightPanelById",
        "summary": "getTemplateForRightPanelById",
        "tags": [
          "Common - Right Panel"
        ],
        "operationId": "getTemplateForRightPanelById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
       "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "ID of template",
            "required": true,
            "type": "string"
            }
            
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/xbuilder/widgets": {
      "post": {
        "description": "",
        "summary": "createwidget",
        "tags": [
          "Widget"
        ],
        "operationId": "createwidget",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createWidgetRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/xbuilder/widgets/{widgetId}": {
      "get": {
        "description": "",
        "summary": "getWidgetById",
        "tags": [
          "Widget"
        ],
        "operationId": "getWidgetById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "widgetId",
            "in": "path",
            "description": "ID of widget to return",
            "required": true,
            "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
},
  "definitions": {
...
    "createWidgetRequest": {
      "title": "createWidgetRequest",
      "example": {
        "name": "serverRoomLine",
        "description": "server Room Line",
        "roles": [
          "DASHBOARD_CONFIGURER",
          "DASHBOARD_MANAGER"
        ],
        "resourceDetails": {
          "sourceId": "SG_1",
          "resourceId": "SG_RD_1",
          "payloadKeyForChart": ""
        },
        "widgetDetails": {
          "chartType": "line",
          "subChartType": "simpleLine",
          "widgetConfig": {
            "title": "room data",
            "titleForX": "room number",
            "unitForX": "NO",
            "dataForX": {
              "payloadKey": "roomNo",
              "dataType": "number"
            },
            "seriesDetails": [
              {
                "id": "SE_1",
                "name": "temperature series",
                "titleForY": "temperature data",
                "unitForY": "C",
                "dataForY": {
                  "payloadKey": "temperature",
                  "dataType": "number"
                }
              }
            ]
          },
          "otherConfig": {}
        },
        "createdBy": "Gabriel Macht"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "resourceDetails": {
          "$ref": "#/definitions/ResourceDetails"
        },
        "widgetDetails": {
          "$ref": "#/definitions/WidgetDetails"
        },
        "createdBy": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description",
        "roles",
        "resourceDetails",
        "widgetDetails",
        "createdBy"
      ]
    },
    "ResourceDetails": {
      "title": "ResourceDetails",
      "example": {
        "sourceId": "SG_1",
        "resourceId": "SG_RD_1",
        "payloadKeyForChart": ""
      },
      "type": "object",
      "properties": {
        "sourceId": {
          "type": "string"
        },
        "resourceId": {
          "type": "string"
        },
        "payloadKeyForChart": {
          "type": "string"
        }
      },
      "required": [
        "sourceId",
        "resourceId",
        "payloadKeyForChart"
      ]
    },
    "WidgetDetails": {
      "title": "WidgetDetails",
      "example": {
        "chartType": "line",
        "subChartType": "simpleLine",
        "widgetConfig": {
          "title": "room data",
          "titleForX": "room number",
          "unitForX": "NO",
          "dataForX": {
            "payloadKey": "roomNo",
            "dataType": "number"
          },
          "seriesDetails": [
            {
              "id": "SE_1",
              "name": "temperature series",
              "titleForY": "temperature data",
              "unitForY": "C",
              "dataForY": {
                "payloadKey": "temperature",
                "dataType": "number"
              }
            }
          ]
        },
        "otherConfig": {}
      },
      "type": "object",
      "properties": {
        "chartType": {
          "type": "string"
        },
        "subChartType": {
          "type": "string"
        },
        "widgetConfig": {
          "$ref": "#/definitions/WidgetConfig"
        },
        "otherConfig": {
          "type": "object"
        }
      },
      "required": [
        "chartType",
        "subChartType",
        "widgetConfig",
        "otherConfig"
      ]
    },
    "WidgetConfig": {
      "title": "WidgetConfig",
      "example": {
        "title": "room data",
        "titleForX": "room number",
        "unitForX": "NO",
        "dataForX": {
          "payloadKey": "roomNo",
          "dataType": "number"
        },
        "seriesDetails": [
          {
            "id": "SE_1",
            "name": "temperature series",
            "titleForY": "temperature data",
            "unitForY": "C",
            "dataForY": {
              "payloadKey": "temperature",
              "dataType": "number"
            }
          }
        ]
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "titleForX": {
          "type": "string"
        },
        "unitForX": {
          "type": "string"
        },
        "dataForX": {
          "$ref": "#/definitions/DataForX"
        },
        "seriesDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SeriesDetail"
          }
        }
      },
      "required": [
        "title",
        "titleForX",
        "unitForX",
        "dataForX",
        "seriesDetails"
      ]
    },
    "DataForX": {
      "title": "DataForX",
      "example": {
        "payloadKey": "roomNo",
        "dataType": "number"
      },
      "type": "object",
      "properties": {
        "payloadKey": {
          "type": "string"
        },
        "dataType": {
          "type": "string"
        }
      },
      "required": [
        "payloadKey",
        "dataType"
      ]
    },
    "SeriesDetail": {
      "title": "SeriesDetail",
      "example": {
        "id": "SE_1",
        "name": "temperature series",
        "titleForY": "temperature data",
        "unitForY": "C",
        "dataForY": {
          "payloadKey": "temperature",
          "dataType": "number"
        }
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "titleForY": {
          "type": "string"
        },
        "unitForY": {
          "type": "string"
        },
        "dataForY": {
          "$ref": "#/definitions/DataForY"
        }
      },
      "required": [
        "id",
        "name",
        "titleForY",
        "unitForY",
        "dataForY"
      ]
    },
    "DataForY": {
      "title": "DataForY",
      "example": {
        "payloadKey": "temperature",
        "dataType": "number"
      },
      "type": "object",
      "properties": {
        "payloadKey": {
          "type": "string"
        },
        "dataType": {
          "type": "string"
        }
      },
      "required": [
        "payloadKey",
        "dataType"
      ]
    },
...
    "createTemplateRequest": {
      "title": "createTemplateRequest",
      "example": {
        "templateName": "Medicinal vehicles",
        "description": "This template is to collect information regarding medicinal vehicles",
        "templateFields": [
          {
            "fieldName": "vehicleName",
            "fieldType": "string",
            "isMandatory": true,
            "isEncrpyted": true
          },
          {
            "fieldName": "vehicleType",
            "fieldType": "string",
            "isMandatory": true,
            "isEncrpyted": false
          },
          {
            "fieldName": "vehicleNumber",
            "fieldType": "string",
            "isMandatory": true,
            "isEncrpyted": true
          },
          {
            "fieldName": "supportEmail",
            "fieldType": "Email",
            "isMandatory": true,
            "isEncrpyted": true
          },
          {
            "fieldName": "supportPhNo",
            "fieldType": "PhoneNumber",
            "country": "India",
            "countryCode": "+91",
            "regex": "^[6789]\\d{9}$",
            "isMandatory": true,
            "isEncrpyted": true
          }
        ],
        "createdBy": "ajayn"
      },
      "type": "object",
      "properties": {
        "templateName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "templateFields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TemplateField"
          }
        },
        "createdBy": {
          "type": "string"
        }
      },
      "required": [
        "templateName",
        "description",
        "templateFields",
        "createdBy"
      ]
    },
    "TemplateField": {
      "title": "TemplateField",
      "example": {
        "fieldName": "vehicleName",
        "fieldType": "string",
        "isMandatory": true,
        "isEncrpyted": true
      },
      "type": "object",
      "properties": {
        "fieldName": {
          "type": "string"
        },
        "fieldType": {
          "type": "string"
        },
        "isMandatory": {
          "type": "boolean"
        },
        "isEncrpyted": {
          "type": "boolean"
        },
        "country": {
          "type": "string"
        },
        "countryCode": {
          "type": "string"
        },
        "regex": {
          "type": "string"
        }
      },
      "required": [
        "fieldName",
        "fieldType",
        "isMandatory",
        "isEncrpyted"
      ]
    },
    "updateTemplateByIdRequest": {
      "title": "updateTemplateByIdRequest",
      "example": {
        "templateName": "Medicinal vehicles gtttt",
        "description": "This template is to collect information regarding medicinal vehicles",
        "templateFields": [
          {
            "fieldName": "vehicleName",
            "fieldType": "string",
            "isMandatory": true,
            "isEncrpyted": true
          },
          {
            "fieldName": "vehicleType",
            "fieldType": "string",
            "isMandatory": true,
            "isEncrpyted": false
          },
          {
            "fieldName": "vehicleNumber",
            "fieldType": "string",
            "isMandatory": true,
            "isEncrpyted": true
          },
          {
            "fieldName": "supportEmail",
            "fieldType": "Email",
            "isMandatory": true,
            "isEncrpyted": true
          },
          {
            "fieldName": "supportPhNo",
            "fieldType": "PhoneNumber",
            "country": "India",
            "countryCode": "+91",
            "regex": "^[6789]\\d{9}$",
            "isMandatory": true,
            "isEncrpyted": true
          }
        ],
        "createdBy": "ajayn",
        "createdOn": "2019-09-27T10:31:41.985Z",
        "modifiedBy": "ajayn"
      },
      "type": "object",
      "properties": {
        "templateName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "templateFields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TemplateField"
          }
        },
        "createdBy": {
          "type": "string"
        },
        "createdOn": {
          "type": "string"
        },
        "modifiedBy": {
          "type": "string"
        }
      },
      "required": [
        "templateName",
        "description",
        "templateFields",
        "createdBy",
        "createdOn",
        "modifiedBy"
      ]
    },
...
  },
...
}